// api/_schema/resolvers-types.ts
export type Maybe<T> = T | undefined;
export type RequireFields<T, K extends keyof T> = {
    [X in Exclude<keyof T, K>]?: T[X];
} &
    { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    DateTime: any;
};

export type CreateUserInput = {
    emailAddress: Scalars['String'];
    password: Scalars['String'];
    firstName: Scalars['String'];
    lastName: Scalars['String'];
    dateOfBirth: Scalars['DateTime'];
    roles?: Maybe<Array<Scalars['String']>>;
};

export enum GrantType {
    Password = 'PASSWORD'
}

export type Mutation = {
    __typename?: 'Mutation';
    _?: Maybe<Scalars['Boolean']>;
    changePassword?: Maybe<UserMutationResponse>;
    createUser?: Maybe<UserMutationResponse>;
    deleteAccount?: Maybe<UserMutationResponse>;
    deleteUser?: Maybe<UserMutationResponse>;
    forgotPassword?: Maybe<MutationResponse>;
    generateToken?: Maybe<Token>;
    lockUser?: Maybe<UserMutationResponse>;
    register?: Maybe<UserMutationResponse>;
    resetPassword?: Maybe<UserMutationResponse>;
    seedData?: Maybe<MutationResponse>;
    unlockUser?: Maybe<UserMutationResponse>;
    updateProfile?: Maybe<UserMutationResponse>;
    updateUser?: Maybe<UserMutationResponse>;
};

export type MutationChangePasswordArgs = {
    currentPassword: Scalars['String'];
    newPassword: Scalars['String'];
};

export type MutationCreateUserArgs = {
    input?: Maybe<CreateUserInput>;
};

export type MutationDeleteUserArgs = {
    id: Scalars['ID'];
};

export type MutationForgotPasswordArgs = {
    emailAddress: Scalars['String'];
};

export type MutationGenerateTokenArgs = {
    input?: Maybe<TokenInput>;
};

export type MutationLockUserArgs = {
    id: Scalars['ID'];
};

export type MutationRegisterArgs = {
    input?: Maybe<RegisterInput>;
};

export type MutationResetPasswordArgs = {
    token: Scalars['String'];
    emailAddress: Scalars['String'];
    newPassword: Scalars['String'];
};

export type MutationUnlockUserArgs = {
    id: Scalars['ID'];
};

export type MutationUpdateProfileArgs = {
    input?: Maybe<UpdateProfileInput>;
};

export type MutationUpdateUserArgs = {
    id: Scalars['ID'];
    input?: Maybe<UpdateUserInput>;
};

export type MutationResponse = {
    __typename?: 'MutationResponse';
    message?: Maybe<Scalars['String']>;
    code?: Maybe<Scalars['String']>;
};

export type Query = {
    __typename?: 'Query';
    _?: Maybe<Scalars['Boolean']>;
    getProfile?: Maybe<User>;
    getUserById?: Maybe<User>;
    searchUsers?: Maybe<UserSearchResult>;
};

export type QueryGetUserByIdArgs = {
    id: Scalars['ID'];
};

export type QuerySearchUsersArgs = {
    input?: Maybe<SearchUserInput>;
};

export type RegisterInput = {
    emailAddress: Scalars['String'];
    password: Scalars['String'];
    firstName: Scalars['String'];
    lastName: Scalars['String'];
    dateOfBirth: Scalars['DateTime'];
};

export type SearchUserInput = {
    size?: Maybe<Scalars['Int']>;
    search?: Maybe<Scalars['String']>;
    sort?: Maybe<UserSortExpression>;
    cursor?: Maybe<Scalars['String']>;
};

export type Token = {
    __typename?: 'Token';
    accessToken: Scalars['String'];
    expiresIn: Scalars['Int'];
};

export type TokenInput = {
    grantType: GrantType;
    emailAddress?: Maybe<Scalars['String']>;
    password?: Maybe<Scalars['String']>;
};

export type UpdateProfileInput = {
    emailAddress: Scalars['String'];
    firstName: Scalars['String'];
    lastName: Scalars['String'];
    dateOfBirth: Scalars['DateTime'];
};

export type UpdateUserInput = {
    emailAddress: Scalars['String'];
    firstName: Scalars['String'];
    lastName: Scalars['String'];
    dateOfBirth: Scalars['DateTime'];
    roles?: Maybe<Array<Scalars['String']>>;
};

export type User = {
    __typename?: 'User';
    id: Scalars['ID'];
    emailAddress: Scalars['String'];
    firstName: Scalars['String'];
    lastName: Scalars['String'];
    dateOfBirth: Scalars['DateTime'];
    isLockedOut?: Maybe<Scalars['Boolean']>;
    roles?: Maybe<Array<Scalars['String']>>;
};

export type UserMutationResponse = {
    __typename?: 'UserMutationResponse';
    message?: Maybe<Scalars['String']>;
    code?: Maybe<Scalars['String']>;
    user?: Maybe<User>;
};

export type UserSearchResult = {
    __typename?: 'UserSearchResult';
    items?: Maybe<Array<User>>;
    before?: Maybe<Scalars['String']>;
    after?: Maybe<Scalars['String']>;
};

export enum UserSortExpression {
    NameAsc = 'NAME_ASC',
    NameDesc = 'NAME_DESC',
    EmailAddressAsc = 'EMAIL_ADDRESS_ASC',
    EmailAddressDesc = 'EMAIL_ADDRESS_DESC'
}

export type WithIndex<TObject> = TObject & Record<string, any>;
export type ResolversObject<TObject> = WithIndex<TObject>;

export type ResolverTypeWrapper<T> = Promise<T> | T;

export type StitchingResolver<TResult, TParent, TContext, TArgs> = {
    fragment: string;
    resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};

export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
    | ResolverFn<TResult, TParent, TContext, TArgs>
    | StitchingResolver<TResult, TParent, TContext, TArgs>;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
    parent: TParent,
    args: TArgs,
    context: TContext,
    info: GraphQLResolveInfo
) => Promise<TResult> | TResult;
// ...
export type ResolversTypes = ResolversObject<{
    Query: ResolverTypeWrapper<{}>;
    Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
    User: ResolverTypeWrapper<User>;
    ID: ResolverTypeWrapper<Scalars['ID']>;
    String: ResolverTypeWrapper<Scalars['String']>;
    DateTime: ResolverTypeWrapper<Scalars['DateTime']>;
    SearchUserInput: SearchUserInput;
    Int: ResolverTypeWrapper<Scalars['Int']>;
    UserSortExpression: UserSortExpression;
    UserSearchResult: ResolverTypeWrapper<UserSearchResult>;
    Mutation: ResolverTypeWrapper<{}>;
    UserMutationResponse: ResolverTypeWrapper<UserMutationResponse>;
    CreateUserInput: CreateUserInput;
    MutationResponse: ResolverTypeWrapper<MutationResponse>;
    TokenInput: TokenInput;
    GrantType: GrantType;
    Token: ResolverTypeWrapper<Token>;
    RegisterInput: RegisterInput;
    UpdateProfileInput: UpdateProfileInput;
    UpdateUserInput: UpdateUserInput;
}>;

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = ResolversObject<{
    Query: {};
    Boolean: Scalars['Boolean'];
    User: User;
    ID: Scalars['ID'];
    String: Scalars['String'];
    DateTime: Scalars['DateTime'];
    SearchUserInput: SearchUserInput;
    Int: Scalars['Int'];
    UserSortExpression: UserSortExpression;
    UserSearchResult: UserSearchResult;
    Mutation: {};
    UserMutationResponse: UserMutationResponse;
    CreateUserInput: CreateUserInput;
    MutationResponse: MutationResponse;
    TokenInput: TokenInput;
    GrantType: GrantType;
    Token: Token;
    RegisterInput: RegisterInput;
    UpdateProfileInput: UpdateProfileInput;
    UpdateUserInput: UpdateUserInput;
}>;

export interface DateTimeScalarConfig
    extends GraphQLScalarTypeConfig<ResolversTypes['DateTime'], any> {
    name: 'DateTime';
}

export type MutationResolvers<
    ContextType = Context,
    ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']
> = ResolversObject<{
    _?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
    changePassword?: Resolver<
        Maybe<ResolversTypes['UserMutationResponse']>,
        ParentType,
        ContextType,
        RequireFields<
            MutationChangePasswordArgs,
            'currentPassword' | 'newPassword'
        >
    >;
    createUser?: Resolver<
        Maybe<ResolversTypes['UserMutationResponse']>,
        ParentType,
        ContextType,
        RequireFields<MutationCreateUserArgs, never>
    >;
    deleteAccount?: Resolver<
        Maybe<ResolversTypes['UserMutationResponse']>,
        ParentType,
        ContextType
    >;
    deleteUser?: Resolver<
        Maybe<ResolversTypes['UserMutationResponse']>,
        ParentType,
        ContextType,
        RequireFields<MutationDeleteUserArgs, 'id'>
    >;
    forgotPassword?: Resolver<
        Maybe<ResolversTypes['MutationResponse']>,
        ParentType,
        ContextType,
        RequireFields<MutationForgotPasswordArgs, 'emailAddress'>
    >;
    generateToken?: Resolver<
        Maybe<ResolversTypes['Token']>,
        ParentType,
        ContextType,
        RequireFields<MutationGenerateTokenArgs, never>
    >;
    lockUser?: Resolver<
        Maybe<ResolversTypes['UserMutationResponse']>,
        ParentType,
        ContextType,
        RequireFields<MutationLockUserArgs, 'id'>
    >;
    register?: Resolver<
        Maybe<ResolversTypes['UserMutationResponse']>,
        ParentType,
        ContextType,
        RequireFields<MutationRegisterArgs, never>
    >;
    resetPassword?: Resolver<
        Maybe<ResolversTypes['UserMutationResponse']>,
        ParentType,
        ContextType,
        RequireFields<
            MutationResetPasswordArgs,
            'token' | 'emailAddress' | 'newPassword'
        >
    >;
    seedData?: Resolver<
        Maybe<ResolversTypes['MutationResponse']>,
        ParentType,
        ContextType
    >;
    unlockUser?: Resolver<
        Maybe<ResolversTypes['UserMutationResponse']>,
        ParentType,
        ContextType,
        RequireFields<MutationUnlockUserArgs, 'id'>
    >;
    updateProfile?: Resolver<
        Maybe<ResolversTypes['UserMutationResponse']>,
        ParentType,
        ContextType,
        RequireFields<MutationUpdateProfileArgs, never>
    >;
    updateUser?: Resolver<
        Maybe<ResolversTypes['UserMutationResponse']>,
        ParentType,
        ContextType,
        RequireFields<MutationUpdateUserArgs, 'id'>
    >;
}>;
// ...
export type QueryResolvers<
    ContextType = Context,
    ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']
> = ResolversObject<{
    _?: Resolver<Maybe<ResolversTypes['Boolean']>, ParentType, ContextType>;
    getProfile?: Resolver<
        Maybe<ResolversTypes['User']>,
        ParentType,
        ContextType
    >;
    getUserById?: Resolver<
        Maybe<ResolversTypes['User']>,
        ParentType,
        ContextType,
        RequireFields<QueryGetUserByIdArgs, 'id'>
    >;
    searchUsers?: Resolver<
        Maybe<ResolversTypes['UserSearchResult']>,
        ParentType,
        ContextType,
        RequireFields<QuerySearchUsersArgs, never>
    >;
}>;